some changes to get less warnings

+++ src/utilities/MoToServo.h
@@ -101,9 +101,9 @@ class MoToServo
     void detach();
     void write(uint16_t);         // specify the angle in degrees, 0 to 180. Values obove 180 are interpreted
                              // as microseconds, limited to MaximumPulse and MinimumPulse
-    void setSpeed(int);      // Set movement speed, the higher the faster
+    void setSpeed(uint16_t);      // Set movement speed, the higher the faster
                              // Zero means no speed control (default)
-    void setSpeed(int,bool); // Set compatibility-Flag (true= compatibility with version V08 and earlier)
+    void setSpeed(uint16_t,bool); // Set compatibility-Flag (true= compatibility with version V08 and earlier)
     #define HIGHRES 0
     #define SPEEDV08 1
 	void setSpeedTime(uint16_t  );  // set Speed as time between 0...180° in milliseconds
